requirements.txt
pytrends>=4.9.2
pandas>=2.0,<3.0
PyYAML>=6.0
